Add files via upload
All of our code, from cleanup to model and UI implementation, was written and run on Google Collab. We therefore recommend using the code present in the file called “BIPxTech_ML_Project_Gr13_.ipynb” on the Collab platform so as not to have problems. The user interface was created to run on Collab and not on other platforms or applications, so we recommend using the interface on Google Collab.